pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agent-messaging"
version = "0.1.0"
description = "Discord surface for local AI CLI agents with persistent sessions and markdown memory."
readme = "docs/plans.md"
requires-python = ">=3.9"
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
discord = [
  "discord.py>=2.5",
]

[project.scripts]
agent-messaging = "agent_messaging.application.app:main"

[tool.setuptools]
package-dir = {"agent_messaging" = "src"}
packages = [
  "agent_messaging",
  "agent_messaging.application",
  "agent_messaging.config",
  "agent_messaging.core",
  "agent_messaging.gateway",
  "agent_messaging.memory",
  "agent_messaging.observability",
  "agent_messaging.providers",
  "agent_messaging.runtime",
  "agent_messaging.services",
]
